test update for MacOS CI
